+++ installer-linux.sh
@@ -1,17 +1,15 @@
 #!/usr/bin/env bash
-cd
-echo "To stop the process, please press CTRL+C"
-echo "And remember, run main.py in the folder"
-echo "Input File Destitnation: "
-read DIR
-if [ ! -d "$DIR" ]; then
-	echo "YOU MUST RESTART TO CONTINUE - DIRECTORY INVALID"
-else
-	cd $DIR
-	mkdir TinyHecticLine
-	cd TinyHecticLine
-	git clone git://github.com/AndrewShen31/Tiny-Hectic-Line
-	py main.py
-fi
-echo "PRESS ENTER TO EXIT"
+mkdir TinyHecticLine
+cd TinyHecticLine
+git clone git://github.com/AndrewShen31/Tiny-Hectic-Line
+cd Tiny-Hectic-Line
+rm -f installer-win.bat
+rm -r -f .git
+rm -r -f .idea
+rm -r -f __pycache__
+clear
+echo Remember, run main.py in the folder
+echo To exit the game, press [ESC] button
+echo PRESS ENTER TO PLAY Tiny-Hectic-Line
 read END
+py main.py